feat(Attach Uploaded Files to PDF): update tests, rename file and update imports
