+++ app/domains/beratungshilfe/services/pdf/index.ts
@@ -12,7 +12,6 @@ import { fillPdf } from "~/services/pdf/fillPdf.server";
 import { createFooter } from "~/services/pdf/footer/createFooter";
 import type { PDFDocumentBuilder } from "~/services/pdf/pdfFromUserData";
 import { pdfFromUserData } from "~/services/pdf/pdfFromUserData";
-import { attachUserUploadedFilesToPdf } from "./attachUserUploadedFilesToPdf";
 import { createChecklistPage } from "./checklist/createChecklistPage";
 import { fillAngelegenheit } from "./pdfForm/A_angelegenheit";
 import { fillVorraussetzungen } from "./pdfForm/B_vorraussetzungen";
@@ -23,6 +22,7 @@ import { fillBesitz } from "./pdfForm/F_besitz/F_besitz";
 import { fillFooter } from "./pdfForm/footer";
 import { fillAusgaben } from "./pdfForm/G_ausgaben";
 import { fillHeader } from "./pdfForm/header";
+import { attachUserUploadedFilesToPdf } from "./userUploadedFiles/attachUserUploadedFilesToPdf";
 
 const METADATA: Metadata = {
   AUTHOR: "Bundesministerium der Justiz",
