+++ app/domains/beratungshilfe/services/pdf/userUploadedFiles/__test__/attachUserUploadedFilesToPdf.test.ts
@@ -1,9 +1,8 @@
 import { PDFDocument } from "pdf-lib";
 import { vi, type Mock } from "vitest";
-import { CheckboxValue } from "~/components/inputs/Checkbox";
 import { type BeratungshilfeFormularContext } from "~/domains/beratungshilfe/formular";
 import { downloadUserFileFromS3 } from "~/services/externalDataStorage/userFileS3Helpers";
-import { attachUserUploadedFilesToPdf } from "..";
+import { attachUserUploadedFilesToPdf } from "../attachUserUploadedFilesToPdf";
 
 vi.mock("~/services/externalDataStorage/userFileS3Helpers", () => ({
   downloadUserFileFromS3: vi.fn(),
@@ -25,7 +24,6 @@ describe("attachUserUploadedFilesToPdf", () => {
     mockMainPDFBuffer = await mainPdfDoc.save();
 
     // Create PDF from user uploaded file
-
     const userPdfDoc = await PDFDocument.create();
     userPdfDoc.addPage();
     mockUserPDFBuffer = await userPdfDoc.save();
@@ -94,72 +92,6 @@ describe("attachUserUploadedFilesToPdf", () => {
     const mockedUserData: BeratungshilfeFormularContext = {
       keineLeistungenBeweis: undefined,
       staatlicheLeistungen: "keine",
-    };
-
-    const resultBuffer = await attachUserUploadedFilesToPdf(
-      mockMainPDFBuffer,
-      sessionId,
-      flowId,
-      mockedUserData,
-    );
-    const resultPdf = await PDFDocument.load(resultBuffer);
-    expect(resultPdf.getPageCount()).toBe(1);
-  });
-
-  it("should attach files only for conditions that are true", async () => {
-    const mockedUserData: BeratungshilfeFormularContext = {
-      keineLeistungenBeweis: [
-        {
-          filename: "testfile1.pdf",
-          fileType: "application/pdf",
-          fileSize: 5938,
-          savedFileKey: "xyz123-lmn456-opq789",
-        },
-      ],
-      wohngeldBeweis: [
-        {
-          filename: "testfile2.pdf",
-          fileType: "application/pdf",
-          fileSize: 19718,
-          savedFileKey: "xyz789-lmn654-opq321",
-        },
-      ],
-      staatlicheLeistungen: "keine",
-      weitereseinkommen: {
-        wohngeld: CheckboxValue.on,
-        unterhaltszahlungen: CheckboxValue.off,
-        arbeitlosengeld: CheckboxValue.off,
-        kindergeld: CheckboxValue.off,
-        bafoeg: CheckboxValue.off,
-        krankengeld: CheckboxValue.off,
-        rente: CheckboxValue.off,
-        elterngeld: CheckboxValue.off,
-        insolvenzgeld: CheckboxValue.off,
-        ueberbrueckungsgeld: CheckboxValue.off,
-        others: CheckboxValue.off,
-      },
-    };
-
-    const resultBuffer = await attachUserUploadedFilesToPdf(
-      mockMainPDFBuffer,
-      sessionId,
-      flowId,
-      mockedUserData,
-    );
-    const resultPdf = await PDFDocument.load(resultBuffer);
-    expect(resultPdf.getPageCount()).toBe(3);
-  });
-  it("should not attach files for conditions that are false", async () => {
-    const mockedUserData: BeratungshilfeFormularContext = {
-      buergergeldBeweis: [
-        {
-          filename: "testfile1.pdf",
-          fileType: "application/pdf",
-          fileSize: 5938,
-          savedFileKey: "xyz123-lmn456-opq789",
-        },
-      ],
-      staatlicheLeistungen: "asylbewerberleistungen",
     };
 
     const resultBuffer = await attachUserUploadedFilesToPdf(
